Update Contour_Stat_Plot_Analysis.R
Removed dotted lines in the plots for comparing VEL_DMP against other methods as these were hard coded based on analysis not in R. 

The dotted lines were intended to indicate whether other methods (e.g. rigid or ECL_DIR) were closer to ground truth than VEL_DMP. The dotted lines are not essential in the plot however, as the manuscript can be revised to comment on this point in manuscript text. 

The dotted line adds confusion to the manuscript and also the code, which should not have any text which is "hard coded" and not traceable in code.